+++ server/services/nuki/lib/mqtt/index.js
@@ -16,7 +16,7 @@ const { subscribeDeviceTopic } = require('./nuki.mqtt.subscribeDeviceTopic');
 const NukiMQTTHandler = function NukiMQTTHandler(nukiHandler) {
   this.nukiHandler = nukiHandler;
   // Gladys MQTT service
-  this.mqttService = null;
+  this.mqttService = this.nukiHandler.gladys.service.getService('mqtt');
   // Found devices
   this.discoveredDevices = {};
 };
