+++ server/test/services/nuki/lib/config/nuki.getConfiguration.test.js
@@ -4,6 +4,7 @@ const sinon = require('sinon');
 const { assert, fake } = sinon;
 const NukiService = require('../../../../../services/nuki/index');
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 
 const gladys = {
   variable: {
@@ -12,6 +13,9 @@ const gladys = {
   },
   device: {
     get: fake.resolves([]),
+  },
+  service: {
+    getService: fake.returns(mqttService),
   },
 };
