+++ server/test/services/nuki/lib/device/nuki.postDelete.test.js
@@ -1,16 +1,11 @@
 const sinon = require('sinon');
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 
 const { assert, fake } = sinon;
 
 const NukiHandler = require('../../../../../services/nuki/lib');
 const NukiMQTTHandler = require('../../../../../services/nuki/lib/mqtt');
-
-const mqttService = {
-  device: {
-    unsubscribe: fake.returns(null),
-  },
-};
 
 const gladys = {
   service: {
