+++ server/test/services/nuki/lib/http/nuki.http.convertToDevice.test.js
@@ -1,10 +1,16 @@
 const sinon = require('sinon');
 const { expect } = require('chai');
+const { fake } = sinon;
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 const NukiHandler = require('../../../../../services/nuki/lib');
 const NukiHTTPHandler = require('../../../../../services/nuki/lib/http');
 
-const gladys = {};
+const gladys = {
+  service: {
+    getService: fake.returns(mqttService),
+  },
+};
 
 describe('nuki.http.convertToDevice', () => {
   let nukiHttpHandler;
