+++ server/test/services/nuki/lib/mqtt/nuki.mqtt.disconnect.test.js
@@ -2,14 +2,10 @@ const sinon = require('sinon');
 
 const { fake } = sinon;
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 const NukiHandler = require('../../../../../services/nuki/lib');
 const NukiMQTTHandler = require('../../../../../services/nuki/lib/mqtt');
 
-const mqttService = {
-  device: {
-    unsubscribe: fake.returns(true),
-  },
-};
 const gladys = {
   service: {
     getService: fake.returns(mqttService),
