+++ server/test/services/nuki/lib/mqtt/nuki.mqtt.scan.test.js
@@ -3,16 +3,9 @@ const sinon = require('sinon');
 
 const { assert, fake } = sinon;
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 const NukiHandler = require('../../../../../services/nuki/lib');
 const NukiMQTTHandler = require('../../../../../services/nuki/lib/mqtt');
-
-const mqttService = {
-  device: {
-    unsubscribe: fake.returns(true),
-    subscribe: fake.returns(true),
-  },
-  isUsed: fake.returns(true),
-};
 
 const gladys = {
   service: {
@@ -27,17 +20,15 @@ describe('nuki.mqtt.scan command', () => {
   let nukiHandler;
 
   beforeEach(() => {
+    sinon.reset();
     const nuki = new NukiHandler(gladys, serviceId);
     nukiHandler = new NukiMQTTHandler(nuki);
     nukiHandler.mqttService = mqttService;
     sinon.spy(nukiHandler, 'handleMessage');
   });
 
-  afterEach(() => {
-    sinon.reset();
-  });
-
   it('should subscribe to mqtt topic', async () => {
+    mqttService.isUsed = fake.resolves(true);
     await nukiHandler.scan();
     assert.callCount(nukiHandler.mqttService.device.unsubscribe, 1);
     nukiHandler.mqttService.device.unsubscribe.firstCall.calledWith('homeassistant/#');
@@ -49,7 +40,7 @@ describe('nuki.mqtt.scan command', () => {
   });
 
   it('should raise an error', async () => {
-    mqttService.isUsed = fake.returns(false);
+    mqttService.isUsed = fake.resolves(false);
     await expect(nukiHandler.scan()).to.be.rejectedWith(Error);
     assert.notCalled(nukiHandler.mqttService.device.unsubscribe);
     assert.notCalled(nukiHandler.mqttService.device.subscribe);
