+++ server/test/services/nuki/lib/mqtt/nuki.mqtt.subscribeDeviceTopic.test.js
@@ -2,14 +2,9 @@ const sinon = require('sinon');
 
 const { assert, fake } = sinon;
 const { serviceId } = require('../../mocks/consts.test');
+const { mqttService } = require('../../mocks/mqtt.mock.test');
 const NukiHandler = require('../../../../../services/nuki/lib');
 const NukiMQTTHandler = require('../../../../../services/nuki/lib/mqtt');
-
-const mqttService = {
-  device: {
-    subscribe: fake.returns(true),
-  },
-};
 
 const gladys = {
   service: {
