+++ server/test/services/nuki/lib/nuki.getDiscoveredDevices.test.js
@@ -1,6 +1,7 @@
 const { expect } = require('chai');
 const sinon = require('sinon');
 
+const { fake } = sinon;
 const NukiHandler = require('../../../../services/nuki/lib');
 
 const { serviceId, existingDevice } = require('../mocks/consts.test');
@@ -13,6 +14,9 @@ const gladys = {
       }
       return undefined;
     },
+  },
+  service: {
+    getService: fake.returns(true),
   },
 };
